requirements.txt
setuptools
PyYAML
torch
pillow
numpy
tqdm
pandas
scikit-learn
threadpoolctl
timm
torchvision
h5py
conch
transformers
opencv-python
scipy
scikit-image
matplotlib
seaborn
yaml